feat(Calendar): add a button to collapse the month and time view
