+++ src/Calendar/test/CalendarSpec.js
@@ -1,23 +1,46 @@
 import React from 'react';
-import { render, fireEvent } from '@testing-library/react';
+import { render } from '@testing-library/react';
+import ReactTestUtils from 'react-dom/test-utils';
 import { parseISO } from '../../utils/dateUtils';
 import { getDOMNode } from '@test/testUtils';
 import { testStandardProps } from '@test/commonCases';
-import Calendar from '../Calendar';
+import CalendarPanel from '../Calendar';
 
 describe('Calendar', () => {
-  testStandardProps(<Calendar />);
+  testStandardProps(<CalendarPanel />);
 
   it('Should render a div with `calendar` class', () => {
-    const instance = getDOMNode(<Calendar calendarDate={new Date(2021, 11, 24)} />);
+    const instance = getDOMNode(<CalendarPanel />);
 
     assert.equal(instance.nodeName, 'DIV');
     assert.ok(instance.className.match(/\bcalendar\b/));
   });
 
+  it('Should be compact', () => {
+    const instance = getDOMNode(<CalendarPanel compact />);
+    assert.ok(instance.className.match(/\bcompact\b/));
+  });
+
+  it('Should be rendered custom elements', () => {
+    const instance = getDOMNode(
+      <CalendarPanel
+        defaultValue={parseISO('2018-07-01')}
+        renderCell={() => {
+          return <i className="text">test</i>;
+        }}
+      />
+    );
+    assert.equal(instance.querySelectorAll('.text').length, 42);
+  });
+
+  it('Should be bordered', () => {
+    const instance = getDOMNode(<CalendarPanel bordered />);
+    assert.ok(instance.className.match(/\bbordered\b/));
+  });
+
   it('Should output valid one day', () => {
     const instance = getDOMNode(
-      <Calendar format="yyyy-MM-dd" calendarDate={parseISO('2018-07-01')} />
+      <CalendarPanel format="yyyy-MM-dd" defaultValue={parseISO('2018-07-01')} />
     );
     assert.equal(
       instance
@@ -27,14 +50,45 @@ describe('Calendar', () => {
     );
   });
 
-  it('Should call `onSelect` callback with the date being clicked', () => {
-    const onSelect = sinon.spy();
+  it('Should call `onSelect` callback', done => {
+    const doneOp = () => {
+      done();
+    };
 
-    const { getByRole } = render(
-      <Calendar format="yyyy-MM-dd" calendarDate={new Date(2021, 11, 24)} onSelect={onSelect} />
+    const instance = getDOMNode(<CalendarPanel format="yyyy-MM-dd" onSelect={doneOp} />);
+
+    ReactTestUtils.Simulate.click(
+      instance.querySelector('.rs-calendar-table-cell-is-today .rs-calendar-table-cell-content')
     );
+  });
 
-    fireEvent.click(getByRole('button', { name: '24 Dec 2021' }));
-    expect(onSelect).to.have.been.calledWith(new Date(2021, 11, 24));
+  it('Should be a controlled value', done => {
+    const instanceRef = React.createRef();
+    const App = React.forwardRef((props, ref) => {
+      const [value, setValue] = React.useState(new Date('6/10/2021'));
+      const pickerRef = React.useRef();
+      React.useImperativeHandle(ref, () => ({
+        panel: pickerRef.current,
+        setDate: date => {
+          setValue(date);
+        }
+      }));
+      return <CalendarPanel value={value} ref={pickerRef} format="yyyy-MM-dd" />;
+    });
+
+    render(<App ref={instanceRef} />);
+    instanceRef.current.setDate(new Date('7/11/2021'));
+    const panel = instanceRef.current.panel;
+
+    assert.equal(panel.querySelector('.rs-calendar-header-title').textContent, 'Jun 2021');
+
+    setTimeout(() => {
+      try {
+        assert.equal(panel.querySelector('.rs-calendar-header-title').textContent, 'Jul 2021');
+        done();
+      } catch (err) {
+        done(err);
+      }
+    }, 100);
   });
 });
