+++ src/DatePicker/DatePicker.tsx
@@ -5,7 +5,7 @@ import pick from 'lodash/pick';
 import omit from 'lodash/omit';
 import IconCalendar from '@rsuite/icons/legacy/Calendar';
 import IconClockO from '@rsuite/icons/legacy/ClockO';
-import { Calendar, CalendarState } from '../Calendar';
+import CalendarContainer from '../Calendar/CalendarContainer';
 import useCalendarDate from '../Calendar/useCalendarDate';
 import Toolbar, { RangeType } from './Toolbar';
 import { DatePickerLocale } from '../locales';
@@ -35,8 +35,6 @@ import {
 } from '../Picker';
 
 import { FormControlBaseProps, PickerBaseProps, RsRefForwardingComponent } from '../@types/common';
-
-import { useCalendarState } from './utils';
 
 export type { RangeType } from './Toolbar';
 
@@ -187,7 +185,7 @@ const DatePicker: RsRefForwardingComponent<'div', DatePickerProps> = React.forwa
       calendarDefaultDate
     );
     const [inputState, setInputState] = useState<InputState>();
-    const { calendarState, reset, openMonth, openTime } = useCalendarState();
+
     const [active, setActive] = useState<boolean>(false);
     const triggerRef = useRef<OverlayTriggerInstance>(null);
     const rootRef = useRef<HTMLDivElement>(null);
@@ -301,32 +299,6 @@ const DatePicker: RsRefForwardingComponent<'div', DatePickerProps> = React.forwa
     );
 
     /**
-     * Toggle month selection panel
-     */
-    const handleMonthDropdown = useCallback(() => {
-      if (calendarState === CalendarState.DROP_MONTH) {
-        reset();
-      } else {
-        openMonth();
-      }
-
-      onToggleMonthDropdown?.(calendarState !== CalendarState.DROP_MONTH);
-    }, [calendarState, onToggleMonthDropdown, openMonth, reset]);
-
-    /**
-     * Switch time selection panel
-     */
-    const handleTimeDropdown = useCallback(() => {
-      if (calendarState === CalendarState.DROP_TIME) {
-        reset();
-      } else {
-        openTime();
-      }
-
-      onToggleTimeDropdown?.(calendarState !== CalendarState.DROP_TIME);
-    }, [calendarState, onToggleTimeDropdown, openTime, reset]);
-
-    /**
      * Callback after clicking the clear button.
      */
     const handleClean = useCallback(
@@ -355,7 +327,7 @@ const DatePicker: RsRefForwardingComponent<'div', DatePickerProps> = React.forwa
       (nextValue: Date, event: React.SyntheticEvent, updatableValue = true) => {
         setCalendarDate(
           // Determine whether the current value contains time, if not, use calendarDate.
-          DateUtils.shouldTime(formatStr)
+          DateUtils.shouldRenderTime(formatStr)
             ? nextValue
             : composeFunctions(
                 (d: Date) => DateUtils.setHours(d, DateUtils.getHours(calendarDate)),
@@ -378,17 +350,17 @@ const DatePicker: RsRefForwardingComponent<'div', DatePickerProps> = React.forwa
     const handleChangePageDate = useCallback(
       (nextPageDate: Date, event: React.MouseEvent) => {
         setCalendarDate(nextPageDate);
-        reset();
         handleDateChange(nextPageDate);
 
         // Show only the calendar month panel. formatStr = 'yyyy-MM'
-        const onlyShowMonth = DateUtils.shouldMonth(formatStr) && !DateUtils.shouldDate(formatStr);
+        const onlyShowMonth =
+          DateUtils.shouldRenderMonth(formatStr) && !DateUtils.shouldRenderDate(formatStr);
 
         if (oneTap && onlyShowMonth) {
           updateValue(event, nextPageDate);
         }
       },
-      [formatStr, handleDateChange, oneTap, reset, setCalendarDate, updateValue]
+      [formatStr, handleDateChange, oneTap, setCalendarDate, updateValue]
     );
 
     const disabledDate = useCallback(
@@ -414,7 +386,7 @@ const DatePicker: RsRefForwardingComponent<'div', DatePickerProps> = React.forwa
         let date = parseDate(value, formatStr);
 
         // If only the time is included in the characters, it will default to today.
-        if (DateUtils.shouldOnlyTime(formatStr)) {
+        if (DateUtils.shouldOnlyRenderTime(formatStr)) {
           date = new Date(`${DateUtils.format(new Date(), 'yyyy-MM-dd')} ${value}`);
         }
 
@@ -453,9 +425,8 @@ const DatePicker: RsRefForwardingComponent<'div', DatePickerProps> = React.forwa
 
     const handleExited = useCallback(() => {
       onClose?.();
-      reset();
       setActive(false);
-    }, [onClose, reset]);
+    }, [onClose]);
 
     // Check whether the time is within the time range of the shortcut option in the toolbar.
     const disabledToolbarHandle = useCallback(
@@ -488,7 +459,7 @@ const DatePicker: RsRefForwardingComponent<'div', DatePickerProps> = React.forwa
     );
 
     const calendar = (
-      <Calendar
+      <CalendarContainer
         {...calendarProps}
         locale={locale}
         showWeekNumbers={showWeekNumbers}
@@ -498,13 +469,12 @@ const DatePicker: RsRefForwardingComponent<'div', DatePickerProps> = React.forwa
         format={formatStr}
         isoWeek={isoWeek}
         inSameMonth={inSameMonth}
-        calendarState={calendarState}
         calendarDate={calendarDate}
         onMoveForward={handleMoveForward}
         onMoveBackward={handleMoveBackward}
         onSelect={handleSelect}
-        onToggleMonthDropdown={handleMonthDropdown}
-        onToggleTimeDropdown={handleTimeDropdown}
+        onToggleMonthDropdown={onToggleMonthDropdown}
+        onToggleTimeDropdown={onToggleTimeDropdown}
         onChangePageDate={handleChangePageDate}
         onChangePageTime={handleChangePageTime}
         onToggleMeridian={handleToggleMeridian}
@@ -558,7 +528,7 @@ const DatePicker: RsRefForwardingComponent<'div', DatePickerProps> = React.forwa
     }, [formatStr, formatDate, placeholder, renderValue, value]);
 
     const caretAs = useMemo(
-      () => caretAsProp || (DateUtils.shouldOnlyTime(formatStr) ? IconClockO : IconCalendar),
+      () => caretAsProp || (DateUtils.shouldOnlyRenderTime(formatStr) ? IconClockO : IconCalendar),
       [caretAsProp, formatStr]
     );
