+++ src/DateRangePicker/styles/index.less
@@ -10,7 +10,7 @@
 .rs-picker-daterange-menu {
   .rs-calendar {
     display: inline-block;
-    height: 276px;
+    height: 274px;
     padding-bottom: 12px;
 
     &:first-child {
