+++ src/utils/dateUtils.ts
@@ -90,14 +90,14 @@ export const omitHideDisabledProps = <T extends Record<string, any>>(
 ): Partial<Omit<T, CalendarOnlyPropsType>> =>
   omitBy<T>(props, (_val, key) => key.startsWith('disabled') || key.startsWith('hide'));
 
-export const shouldTime = (format: string) => /([Hhms])/.test(format);
+export const shouldRenderTime = (format: string) => /([Hhms])/.test(format);
 
-export const shouldMonth = (format: string) => /[Yy]/.test(format) && /[ML]/.test(format);
+export const shouldRenderMonth = (format: string) => /[Yy]/.test(format) && /[ML]/.test(format);
 
-export const shouldDate = (format: string): boolean =>
+export const shouldRenderDate = (format: string): boolean =>
   /[Yy]/.test(format) && /[ML]/.test(format) && /[Dd]/.test(format); // for date-fns v1 and v2
 
-export const shouldOnlyTime = (format: string) =>
+export const shouldOnlyRenderTime = (format: string) =>
   /([Hhms])/.test(format) && !/([YyMDd])/.test(format); // for date-fns v1 and v2
 
 /**
